refactor(core): extract createSelectionSourceSync
Pull the targetElement -> resolveSource -> actions.setSelectionSource
effect out of init() into a focused sync. The factory takes the
targetElement accessor and the setSelectionSource action, owns the
stale-request version counter internally, and runs its own effect.

Dropped ~30 lines from init() and unified the three nearly-identical
`if (requestVersion === currentVersion) ...` guards on the success,
clear, and error paths into a single `clearSource` helper used in
both the then() and catch() branches.

index.tsx: 3048 -> 3018 (-30).
